Made index.js the smart component and ResultsPage pure rendering
